Move online inference into a "separate" package.
