+++ ml_project/tests/features/test_build_features.py
@@ -17,7 +17,7 @@ from heart_disease.features.build_features import \
     serialize_pipeline, \
     deserialize_pipeline, \
     extract_target
-from tests.conftest import get_feature_config
+from conftest import get_feature_config
 
 EPS = 1e-6
