+++ online_inference/requirements.txt
@@ -1,0 +1,6 @@
+pandas==1.2.4
+scikit-learn==0.24.1
+pytest==6.2.3
+uvicorn[standard]==0.13.4
+fastapi==0.64.0
+python-dotenv==0.17.1
